looter is now able to find and read monster loot windows

+++ sources/inc/graphics/Image.hpp
@@ -113,7 +113,8 @@ namespace Amb
 
                 img.pixels.reserve(reqw*reqh);
 
-                Pos base = { reqx, reqy + reqh - 1 };
+                Pos base = { static_cast<int>(reqx), 
+                             static_cast<int>(reqy + reqh - 1) };
 
                 auto it = img.pixels.begin();
                 for (size_t y = 0; y < reqh; ++y)
